refactoring

+++ integration/integration.domain/src/test/java/de/mq/analysis/integration/support/RealFunctionTest.java
@@ -1,10 +1,11 @@
 package de.mq.analysis.integration.support;
 
+import org.junit.Assert;
 import org.junit.Test;
 
 import de.mq.analysis.integration.RealFunction;
 import de.mq.analysis.integration.support.RealFunctionImpl;
-import junit.framework.Assert;
+
 
 public class RealFunctionTest {
 	
@@ -12,12 +13,12 @@ public class RealFunctionTest {
 	
 	@Test
 	public final void f() {
-		Assert.assertEquals(1d/Math.E, realFunction.f(1d));
+		Assert.assertEquals(Double.valueOf(1d/Math.E), Double.valueOf(realFunction.f(1d)));
 	}
 	
 	@Test
 	public final void f0(){
-		Assert.assertEquals(1d, realFunction.f(0));
+		Assert.assertEquals(Double.valueOf(1d), Double.valueOf(realFunction.f(0)));
 	}
 
 }
